Refactor set aria function.
